requirements.txt
# Core ML
torch>=2.1.0,<3
transformers>=4.38.0,<5
accelerate>=0.27.0

# Vector search
faiss-cpu>=1.8.0,<2

# Images & UI
Pillow>=10.2.0
gradio>=4.44.0,<6

# Utilities
numpy>=1.26.0,<3
